separate header and navigation components
