+++ src/components/others/navigation/index.module.scss
@@ -1,0 +1,147 @@
+@use "src/stylesheets/variables/typo";
+@use "src/stylesheets/variables/spacings";
+
+$breakpoint-tablet: 780px;
+$breakpoint-mobile: 480px;
+$breakpoint-desktop: 1200px;
+$body-margin: 2.5rem;
+
+@mixin hideBelowTabletSize {
+  @media only screen and (max-width: $breakpoint-tablet) {
+    display: none;
+  }
+}
+
+.navWrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  height: 3.5rem;
+
+  .nav {
+    @include hideBelowTabletSize();
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+
+    a {
+      color: #333;
+      text-decoration: none;
+      margin-right: 20px;
+      font-size: typo.$s;
+
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+
+    h2 {
+      display: none;
+    }
+  }
+
+  .navEthereum {
+    display: flex;
+    flex-direction: row;
+  }
+}
+
+.overlay {
+  position: fixed;
+  left: 0;
+  right: 0;
+  height: 100%;
+  max-height: 100vh;
+  background-color: hsl(var(--starbright-600));
+  background-repeat: repeat;
+  z-index: 999;
+  transition: opacity 0.4s ease-out;
+  opacity: 0;
+  visibility: hidden;
+
+  @media screen and (max-width: $breakpoint-tablet) {
+    top: 260px;
+  }
+
+  @media screen and (max-width: $breakpoint-mobile) {
+    top: 320px;
+  }
+
+  &.visible {
+    opacity: 1;
+    visibility: visible;
+  }
+}
+
+@keyframes fadeIn {
+  0% {
+    opacity: 0;
+    visibility: hidden;
+  }
+  100% {
+    opacity: 1;
+    visibility: visible;
+  }
+}
+
+.overlayNav {
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  align-items: center;
+  height: 100px;
+  margin-top: spacings.$s-12;
+
+  a {
+    visibility: hidden;
+  }
+
+  &.visible {
+    a {
+      color: #333;
+      text-decoration: none;
+      margin-bottom: 20px;
+      font-size: typo.$l;
+
+      animation: fadeIn 0.3s ease-out;
+      animation-fill-mode: forwards;
+
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+
+    a:nth-child(1) {
+      animation-delay: 0.2s;
+    }
+
+    a:nth-child(2) {
+      animation-delay: 0.3s;
+    }
+
+    a:nth-child(3) {
+      animation-delay: 0.4s;
+    }
+
+    a:nth-child(4) {
+      animation-delay: 0.5s;
+    }
+
+    a:nth-child(5) {
+      animation-delay: 0.6s;
+    }
+
+    a:nth-child(6) {
+      animation-delay: 0.7s;
+    }
+  }
+
+  h2 {
+    font-size: typo.$l;
+    font-weight: bold;
+    margin: 0 0 20px;
+  }
+
+}
